Give establishments.cep a real foreign key, and trim three more tables
The CEP link finally becomes a constraint the database enforces. Two
obstacles were cleared earlier — unmatched CEPs now become NULL, and the
CEP import upserts instead of deleting — leaving only a type mismatch:
Postgres needs comparable types for a foreign key, and there is no
`integer = varchar` operator. So `correios_cep.cep` and
`cep_coordenadas.cep` become INTEGER, matching `establishments.cep`. That
also drops five bytes a row from both and removes the cast from the
proximity join.

`correios_cep` is an ordinary model now. It was excluded from Alembic
back when edne-correios-loader owned the schema and rebuilt the table on
every import; today the library only ever populates the scratch table, so
keeping the exclusion just meant nothing could reference it. `env.py` now
excludes the scratch table instead, which is the thing that really is
transient.

Three further savings, all in tables with tens of millions of rows:

  * `estabelecimentos_staging.cnpj_basico` was the CNPJ with its last four
    positions removed, which in base 36 is an integer division by 36^4.
    Storing eight bytes a row to repeat what `cnpj` already holds cost
    ~500MB across ~63M rows; the build's join computes it instead.
  * `simples_staging` no longer stores rows saying "neither Simples nor
    MEI". The build LEFT JOINs with `coalesce(..., false)`, so an absent
    row means exactly the same thing — and the file lists everyone who
    ever opted in, including those who left.
  * `socios.id` fits in an Integer at ~24M rows. The TRUNCATE opening the
    group now does RESTART IDENTITY, without which the sequence would
    climb across imports until it overflowed.

While in there: dropped the ProgrammingError handlers in the address
router that existed only because `correios_cep` might not exist yet. It
is migration-managed now, so swallowing that error would only hide a real
SQL fault behind an empty list. Also removed a stray f-prefix on a string
with no placeholders.

Verified end to end: the foreign key rejects an establishment whose CEP is
absent from `correios_cep` and blocks deleting a referenced CEP; the
address endpoints still return CEPs zero-padded to eight positions; the
proximity join works across the new integer types; and the build's joins
still resolve company name, legal nature and Simples/MEI correctly with
the derived root and the thinned Simples table.

+++ app/importer/osm_ceps.py
@@ -149,7 +149,9 @@ def import_ceps_from_osm(db: Session) -> int:
 
         result = db.execute(text("""
             INSERT INTO cep_coordenadas (cep, latitude, longitude, source, updated_at)
-            SELECT cep, latitude::numeric, longitude::numeric, 'osm_extract', :now
+            -- cep é INTEGER na tabela (casa com correios_cep.cep); a temp
+            -- recebe texto direto do COPY e o cast sai aqui.
+            SELECT cep::integer, latitude::numeric, longitude::numeric, 'osm_extract', :now
             FROM tmp_osm_ceps
             ON CONFLICT (cep) DO NOTHING
         """), {"now": datetime.now(timezone.utc)})
